refactor(study): address phase7 memory review
