+++ plugin/plugins/study_companion/surfaces/memory_importer.tsx
@@ -1,54 +1,12 @@
 import { useEffect, useState } from '@neko/plugin-ui';
 import type { PluginSurfaceProps } from '@neko/plugin-ui';
+import { callPlugin, errorMessage, text } from './memory_shared';
 
 type MemoryDeck = {
   id: string;
   name: string;
   deck_type: string;
 };
-
-async function readJsonResponse(response: Response, label: string) {
-  if (!response.ok) {
-    throw new Error(`${label} failed: HTTP ${response.status}`);
-  }
-  return await response.json();
-}
-
-async function callPlugin(entryId: string, args: Record<string, unknown> = {}) {
-  const created = await readJsonResponse(await fetch('/runs', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ plugin_id: 'study_companion', entry_id: entryId, args }),
-  }), 'Run create');
-  const runId = created.run_id || created.id;
-  if (!runId) {
-    throw new Error('Run id missing');
-  }
-  for (let attempt = 0; attempt < 40; attempt += 1) {
-    await new Promise((resolve) => window.setTimeout(resolve, 350));
-    const run = await readJsonResponse(await fetch(`/runs/${runId}`), 'Run poll');
-    if (run.status === 'succeeded') {
-      const exported = await readJsonResponse(await fetch(`/runs/${runId}/export`), 'Run export');
-      const item = (exported.items || []).find((candidate: any) => candidate.type === 'json' && candidate.json);
-      if (!item) {
-        throw new Error('Run export missing JSON result');
-      }
-      if (item.json.success === false || item.json.error) {
-        throw new Error(item.json.error?.message || item.json.message || 'Plugin call failed');
-      }
-      return item.json.data || {};
-    }
-    if (['failed', 'canceled', 'timeout'].includes(run.status)) {
-      throw new Error(run.error?.message || run.message || run.status);
-    }
-  }
-  throw new Error('Plugin call timed out');
-}
-
-function text(props: PluginSurfaceProps, key: string, fallback: string) {
-  const value = props.t?.(key);
-  return value && value !== key ? value : fallback;
-}
 
 export default function MemoryImporter(props: PluginSurfaceProps) {
   const [decks, setDecks] = useState<MemoryDeck[]>([]);
@@ -59,13 +17,19 @@ export default function MemoryImporter(props: PluginSurfaceProps) {
   const [busy, setBusy] = useState(false);
 
   useEffect(() => {
-    callPlugin('study_memory_list_decks', { limit: 100 })
-      .then((payload: any) => {
+    const controller = new AbortController();
+    callPlugin<{ decks?: MemoryDeck[] }>('study_memory_list_decks', { limit: 100 }, controller.signal)
+      .then((payload) => {
         const nextDecks = Array.isArray(payload.decks) ? payload.decks : [];
         setDecks(nextDecks);
         setDeckId(nextDecks[0]?.id || '');
       })
-      .catch((error) => setResult(error instanceof Error ? error.message : String(error)));
+      .catch((error) => {
+        if (!controller.signal.aborted) {
+          setResult(errorMessage(error));
+        }
+      });
+    return () => controller.abort();
   }, []);
 
   async function importWords() {
@@ -75,10 +39,10 @@ export default function MemoryImporter(props: PluginSurfaceProps) {
     }
     setBusy(true);
     try {
-      const payload = await callPlugin('study_memory_import_words', { deck_id: deckId, content, fmt });
+      const payload = await callPlugin<Record<string, unknown>>('study_memory_import_words', { deck_id: deckId, content, fmt });
       setResult(JSON.stringify(payload, null, 2));
     } catch (error) {
-      setResult(error instanceof Error ? error.message : String(error));
+      setResult(errorMessage(error));
     } finally {
       setBusy(false);
     }
