+++ plugin/plugins/study_companion/surfaces/passage_recitation.tsx
@@ -1,48 +1,10 @@
 import { useState } from '@neko/plugin-ui';
 import type { PluginSurfaceProps } from '@neko/plugin-ui';
+import { callPlugin, errorMessage, text } from './memory_shared';
 
-async function readJsonResponse(response: Response, label: string) {
-  if (!response.ok) {
-    throw new Error(`${label} failed: HTTP ${response.status}`);
-  }
-  return await response.json();
-}
-
-async function callPlugin(entryId: string, args: Record<string, unknown> = {}) {
-  const created = await readJsonResponse(await fetch('/runs', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ plugin_id: 'study_companion', entry_id: entryId, args }),
-  }), 'Run create');
-  const runId = created.run_id || created.id;
-  if (!runId) {
-    throw new Error('Run id missing');
-  }
-  for (let attempt = 0; attempt < 40; attempt += 1) {
-    await new Promise((resolve) => window.setTimeout(resolve, 350));
-    const run = await readJsonResponse(await fetch(`/runs/${runId}`), 'Run poll');
-    if (run.status === 'succeeded') {
-      const exported = await readJsonResponse(await fetch(`/runs/${runId}/export`), 'Run export');
-      const item = (exported.items || []).find((candidate: any) => candidate.type === 'json' && candidate.json);
-      if (!item) {
-        throw new Error('Run export missing JSON result');
-      }
-      if (item.json.success === false || item.json.error) {
-        throw new Error(item.json.error?.message || item.json.message || 'Plugin call failed');
-      }
-      return item.json.data || {};
-    }
-    if (['failed', 'canceled', 'timeout'].includes(run.status)) {
-      throw new Error(run.error?.message || run.message || run.status);
-    }
-  }
-  throw new Error('Plugin call timed out');
-}
-
-function text(props: PluginSurfaceProps, key: string, fallback: string) {
-  const value = props.t?.(key);
-  return value && value !== key ? value : fallback;
-}
+type RecitationPayload = {
+  diff?: unknown;
+};
 
 export default function PassageRecitation(props: PluginSurfaceProps) {
   const [itemId, setItemId] = useState('');
@@ -58,14 +20,14 @@ export default function PassageRecitation(props: PluginSurfaceProps) {
     }
     setBusy(true);
     try {
-      const payload = await callPlugin('study_memory_recitation_attempt', {
+      const payload = await callPlugin<RecitationPayload>('study_memory_recitation_attempt', {
         item_id: itemId,
         user_input_text: userInput,
         hint_count: hintCount,
       });
       setResult(JSON.stringify(payload.diff || payload, null, 2));
     } catch (error) {
-      setResult(error instanceof Error ? error.message : String(error));
+      setResult(errorMessage(error));
     } finally {
       setBusy(false);
     }
